improve pool modularization
