+++ roles/pool/src/lib/downstream/message_handler.rs
@@ -1,4 +1,4 @@
-use super::super::mining_pool::Downstream;
+use super::Downstream;
 use roles_logic_sv2::{
     errors::Error,
     handlers::mining::{ParseDownstreamMiningMessages, SendTo, SupportedChannelTypes},
