+++ roles/pool/src/lib/mod.rs
@@ -2,3 +2,5 @@ pub mod error;
 pub mod mining_pool;
 pub mod status;
 pub mod template_receiver;
+
+pub mod downstream;
